Move all interfaces to one file
