+++ src/client/src/pages/detail/form/stratigraphy/components/lithologyTable/useLithologyTableState.test.ts
@@ -1,9 +1,7 @@
 // @vitest-environment jsdom
 import { act, renderHook } from "@testing-library/react";
 import { describe, expect, it } from "vitest";
-import { FaciesDescription } from "../../faciesDescription.ts";
-import { LithologicalDescription } from "../../lithologicalDescription.ts";
-import { Lithology } from "../../lithology.ts";
+import { FaciesDescription, LithologicalDescription, Lithology } from "../../stratigraphy.ts";
 import { LithologyTableStateOptions, useLithologyTableState } from "./useLithologyTableState.ts";
 
 const lithology = (overrides: Partial<Lithology> = {}): Lithology => ({
