+++ src/client/src/pages/detail/form/stratigraphy/components/lithologyTable/useLithologyTableState.ts
@@ -1,9 +1,6 @@
 import { useMemo, useRef, useState } from "react";
 import { v4 as uuidv4 } from "uuid";
-import { FaciesDescription } from "../../faciesDescription.ts";
-import { LithologicalDescription } from "../../lithologicalDescription.ts";
-import { Lithology } from "../../lithology.ts";
-import { BaseLayer, DepthLayer } from "../../stratigraphy.ts";
+import { BaseLayer, DepthLayer, FaciesDescription, LithologicalDescription, Lithology } from "../../stratigraphy.ts";
 import {
   createEmptyLithology,
   flagErrors,
