+++ src/client/src/pages/detail/form/stratigraphy/faciesDescription.ts
@@ -1,14 +1,7 @@
 import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
 import { fetchApiV2WithApiError } from "../../../../api/fetchApiV2.ts";
-import { Codelist } from "../../../../components/codelist.ts";
 import { useResetTabStatus } from "../../../../hooks/useResetTabStatus.ts";
-import { BaseLayer } from "./stratigraphy.ts";
-
-export interface FaciesDescription extends BaseLayer {
-  description?: string;
-  faciesId?: number | null;
-  facies?: Codelist | null;
-}
+import { FaciesDescription } from "./stratigraphy.ts";
 
 const fetchFaciesDescriptionsByProfileId = async (profileId: number): Promise<FaciesDescription[]> => {
   return await fetchApiV2WithApiError<FaciesDescription[]>(`faciesdescription?stratigraphyId=${profileId}`, "GET");
