+++ src/client/src/pages/detail/form/stratigraphy/lithology/form/lithologyConsolidatedForm.tsx
@@ -3,7 +3,7 @@ import { useTranslation } from "react-i18next";
 import { Divider } from "@mui/material";
 import { BoreholesCard } from "../../../../../../components/boreholesCard.tsx";
 import { FormContainer, FormDomainMultiSelect, FormDomainSelect } from "../../../../../../components/form/form.ts";
-import { LithologyDescriptionEditForm, LithologyEditForm } from "../../lithology.ts";
+import { LithologyDescriptionEditForm, LithologyEditForm } from "../../stratigraphy.ts";
 import { useLithologyDescriptionShareSync } from "../useLithologyDescriptionShareSync.ts";
 import { LithologyDescriptionForm } from "./lithologyDescriptionForm.tsx";
