+++ src/client/src/pages/detail/form/stratigraphy/lithology/form/lithologyDescriptionForm.tsx
@@ -1,7 +1,7 @@
 import { FC, ReactNode } from "react";
 import { UseFormReturn } from "react-hook-form";
 import { FormCheckbox, FormContainer, FormInput, FormValueType } from "../../../../../../components/form/form.ts";
-import { Lithology } from "../../lithology.ts";
+import { Lithology } from "../../stratigraphy.ts";
 
 interface LithologyDescriptionFormProps {
   lithologyId: number;
