+++ src/client/src/pages/detail/form/stratigraphy/lithology/lithologyLabels.tsx
@@ -3,7 +3,7 @@ import { useTranslation } from "react-i18next";
 import { Typography } from "@mui/material";
 import type { TFunction } from "i18next";
 import { useCodelistDisplayValues } from "../../../../../components/codelist.ts";
-import { Lithology, LithologyDescription } from "../lithology";
+import { Lithology, LithologyDescription } from "../stratigraphy.ts";
 
 const uselessStrings = new Set(["keine Angabe", "sans indication", "senza indicazioni", "not specified"]);
