+++ src/client/src/pages/detail/form/stratigraphy/stratigraphy.ts
@@ -1,9 +1,11 @@
+import { UseFormReturn } from "react-hook-form";
 import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
 import { NullableDateString, User } from "../../../../api/apiInterfaces.ts";
 import { boreholeQueryKey, BoreholeV2 } from "../../../../api/borehole.ts";
+import { ExtractionBoundingBox } from "../../../../api/dataextractionInterfaces.ts";
 import { fetchApiV2WithApiError } from "../../../../api/fetchApiV2.ts";
+import { Codelist } from "../../../../components/codelist.ts";
 import { useResetTabStatus } from "../../../../hooks/useResetTabStatus.ts";
-import { Lithology } from "./lithology.ts";
 
 export interface Stratigraphy {
   id: number;
@@ -47,6 +49,121 @@ export interface DepthLayer {
   toDepth: number | null;
   hasFromDepthError?: boolean;
   hasToDepthError?: boolean;
+}
+
+export interface LithologyDescription {
+  id: number;
+  lithologyId: number;
+  lithology?: Lithology;
+  isFirst: boolean;
+  colorPrimaryId?: number | null;
+  colorPrimary?: Codelist;
+  colorSecondaryId?: number | null;
+  colorSecondary?: Codelist;
+  lithologyUnconMainId?: number | null;
+  lithologyUnconMain?: Codelist;
+  lithologyUncon2Id?: number | null;
+  lithologyUncon2?: Codelist;
+  lithologyUncon3Id?: number | null;
+  lithologyUncon3?: Codelist;
+  lithologyUncon4Id?: number | null;
+  lithologyUncon4?: Codelist;
+  lithologyUncon5Id?: number | null;
+  lithologyUncon5?: Codelist;
+  lithologyUncon6Id?: number | null;
+  lithologyUncon6?: Codelist;
+  componentUnconOrganicCodelistIds?: number[];
+  componentUnconOrganicCodelists?: Codelist[];
+  componentUnconDebrisCodelistIds?: number[];
+  componentUnconDebrisCodelists?: Codelist[];
+  grainShapeCodelistIds?: number[];
+  grainShapeCodelists?: Codelist[];
+  grainAngularityCodelistIds?: number[];
+  grainAngularityCodelists?: Codelist[];
+  hasStriae?: boolean;
+  lithologyUnconDebrisCodelistIds?: number[];
+  lithologyUnconDebrisCodelists?: Codelist[];
+  lithologyConId?: number | null;
+  lithologyCon?: Codelist;
+  componentConParticleCodelistIds?: number[];
+  componentConParticleCodelists?: Codelist[];
+  componentConMineralCodelistIds?: number[];
+  componentConMineralCodelists?: Codelist[];
+  grainSizeId?: number | null;
+  grainSize?: Codelist;
+  grainAngularityId?: number | null;
+  grainAngularity?: Codelist;
+  gradationId?: number | null;
+  gradation?: Codelist;
+  cementationId?: number | null;
+  cementation?: Codelist;
+  structureSynGenCodelistIds?: number[];
+  structureSynGenCodelists?: Codelist[];
+  structurePostGenCodelistIds?: number[];
+  structurePostGenCodelists?: Codelist[];
+}
+
+export interface Lithology extends BaseLayer {
+  isUnconsolidated: boolean | null;
+  hasBedding: boolean;
+  share?: number;
+  shareInverse?: number;
+  lithologyDescriptions?: LithologyDescription[];
+  alterationDegreeId?: number | null;
+  alterationDegree?: Codelist;
+  notes?: string;
+  compactnessId?: number | null;
+  compactness?: Codelist;
+  cohesionId?: number | null;
+  cohesion?: Codelist;
+  humidityId?: number | null;
+  humidity?: Codelist;
+  consistencyId?: number | null;
+  consistency?: Codelist;
+  plasticityId?: number | null;
+  plasticity?: Codelist;
+  uscsTypeCodelistIds?: number[];
+  uscsTypeCodelists?: Codelist[];
+  uscsDeterminationId?: number | null;
+  uscsDetermination?: Codelist;
+  rockConditionCodelistIds?: number[];
+  rockConditionCodelists?: Codelist[];
+  textureMetaCodelistIds?: number[];
+  textureMetaCodelists?: Codelist[];
+}
+
+export interface LayerDepth {
+  fromDepth: number;
+  toDepth: number;
+  lithologyId: number;
+  hasFromDepthError?: boolean;
+  hasToDepthError?: boolean;
+}
+
+export interface LithologyEditForm {
+  lithologyId: number;
+  formMethods: UseFormReturn<Lithology>;
+}
+
+export interface LithologyDescriptionEditForm extends LithologyEditForm {
+  isFirst: boolean;
+  hasBedding?: boolean;
+}
+
+export interface LithologicalDescription extends BaseLayer {
+  description?: string;
+}
+
+export interface ExtractedLithologicalDescription extends LithologicalDescription {
+  startDepthBoundingBoxes: ExtractionBoundingBox[];
+  endDepthBoundingBoxes: ExtractionBoundingBox[];
+  descriptionBoundingBoxes: ExtractionBoundingBox[];
+}
+
+export interface FaciesDescription extends BaseLayer {
+  description?: string;
+  faciesId?: number | null;
+  facies?: Codelist | null;
 }
 
 interface Chronostratigraphy {
